Change all attributes with type double to float

+++ tim7iss/src/main/java/com/tim7/iss/tim7iss/DTOs/Member2/LocationDTOs/LocationResponseDTO.java
@@ -10,8 +10,8 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class LocationResponseDTO {
     String address;
-    Double latitude;
-    Double longitude;
+    Float latitude;
+    Float longitude;
     public LocationResponseDTO(Location location){
         this.address = location.getName();
         this.longitude = location.getLongitude();
